fix(infra): let ScheduleModule and RedisModule forRootAsync take imports
Seven of the nine configured modules take an `AsyncModuleConfig`. These two took a
bare `FactoryProvider` and forwarded nothing, so a factory injecting a provider from
a non-global module was a boot error there while working everywhere else. The
documented case survived only because `ConfigModule.forRoot` is `global: true`.

`AsyncModuleConfig`'s own doc comment says why it exists: a dynamic module is its
own scope, so the module that exports the token has to be in *that* module's
imports, and importing it alongside does not reach the factory. `RedisModule`
forwards them to a named connection too.

Found by writing the `forRootAsync` tests these two never had, and the first version
passed against the bug: an unbound class self-binds into whichever scope asks first,
so a plain class resolved with or without the fix. Both tests use a `token()`
instead, and each was checked to fail with the forwarding removed.

`schedule.test.ts` crossed the 500-line cap, so the module tests are their own file
and `Quiet` moved to a shared `quiet.fixture.ts` rather than being copied.
